Import calibration load script
This script uses the serial number of the attached kinect to identify the turtlebot that is currently connected and attempts to load a calibration file for it from a standard location.
Needs more work to detect an xtion plugged in instead of a kinect, as well as if neither are plugged in.