Attempt to use the usb to serial converter's id
If the robot receives a velocity command we'll assume that it's been
brought up and must now move.  If we don't have the kinect's serial number
yet assume we won't ever get it.  Instead attempt to find a calibration file
that matches the serial number of the usb to serial converter that should be
attached at /dev/ttyUSB0.  This setup will prevent the calibration script from
running indefinitely waiting for a kinect that might never come.